Calling sf::shader::bind rather than glUseProgramObjectARB as the latter causes linker errors
For some unknown reason, calling any function from glew that is defined
with GLEW_GET_FUN() causes a relocation in .text when trying to link
the final executable together. bind(NULL) is called on the shader here
because, as seen in older code revisions of typhon,
sf::shader::unbind(deprecated) was called.