Fixed linker error
It used to be that if you called any glew function pointer you would end
up with a linker error when trying to compile typhon. This was caused
simply because glew wasn't included(who would've thought?). Reverted
from sf:Shader::bind(NULL) to glUseProgram(0) just for the sake of being
1-1 as much as possible with the original project until I can get it
working.