Fixed freeze on getShader(id).setUniform () and added a new class
As stated by the title, this change brings a new class to typhon,
ShaderManager.This was originally part of an attempt to fix a bug
which caused typhon to freeze or sometimes even segfault when trying to
make any call to sf::Shader::setUniform with a shader from getShader().
The cause was (most likely) memory layout confusion between parts of
typhon code, however I never managed to prove it in any way.

**Typhon should now run in all modes.** However, while the shaders load fine and
are valid, they don't show up on the screen. This may be linked to
another issue with texture loading which I haven't gotten around to fix
yet.

Stay tuned for more changes(new classes and architectural changes in
general), as there may be more bugs whose (best) solution is to make
such changes. I'll discuss more about this when the time comes.