Switch to localhost ip instead of 0.0.0.0
A while ago, I removed some URLs from the default config and swapped
them with 0.0.0.0 instead. However, the definition of 0.0.0.0 tends to
be different based on a few factors, so I chose to instead switch to the
address of the loopback interface which is the same on all systems to
avoid any unexpected behaviour.