Added a .gitignore in the build directory so it can be commited to the repo
Does basically what the title says. As an added bonus, the .gitignore is
configured as such that is should ignore every other file that exists in
this directory, so you should be able to build in it without worrying
about your build output showing up on github when you make a commit.