Allow framerate to be set via config
Typhon's framerate was originally hardcoded to be 30 fps.
This commit now allows, as the title says, to set your desired framerate
via a config file, specifically typhonrc in this case. The value is
"framerate", under the "globalints" map. As an added bonus, the default
framerate is now 60 fps. This also paves the way for setting the
framerate from inside typhon itself(not implemented yet) which is the
final goal.